+++ tests/unit/mocked/replicator_test.py
@@ -23,13 +23,14 @@ import unittest
 import mock
 import requests
 
+from cloudant.database import CouchDatabase
 from cloudant.errors import CloudantException
-from cloudant.replicator import ReplicatorDatabase
+from cloudant.replicator import Replicator
 from cloudant.document import Document
 
-class ReplicatorDatabaseTests(unittest.TestCase):
+class ReplicatorTests(unittest.TestCase):
     """
-    Tests for ReplicatorDatabase class
+    Tests for Replicator class
 
     """
     def setUp(self):
@@ -50,19 +51,23 @@ class ReplicatorDatabaseTests(unittest.TestCase):
         self.username = "steve"
         self.password = "abc123"
 
+        self.mock_account = mock.Mock()
+        self.mock_account.__getitem__ = mock.Mock()
+        self.mock_account.__getitem__.return_value = CouchDatabase(
+            self.mock_account,
+            '_replicator'
+        )
+        self.mock_account.session = mock.Mock()
+        self.mock_account.session.return_value = {
+            "userCtx": "user Context"
+        }
+
     def tearDown(self):
         self.patcher.stop()
 
     def test_create_replication(self):
         """test create_replication method"""
-        with mock.patch('cloudant.database.CloudantDatabase.create_document') as mock_create:
-
-            mock_account = mock.Mock()
-            mock_account.session = mock.Mock()
-            mock_account.session.return_value = {
-                "userCtx": "user Context"
-            }
-
+        with mock.patch('cloudant.database.CouchDatabase.create_document') as mock_create:
             mock_target = mock.Mock()
             mock_target.database_url = "http://bob.cloudant.com/target"
             mock_target.creds = {'basic_auth': "target_auth"}
@@ -70,10 +75,8 @@ class ReplicatorDatabaseTests(unittest.TestCase):
             mock_source.database_url = "http://bob.cloudant.com/source"
             mock_source.creds = {'basic_auth': "source_auth"}
 
-            repl = ReplicatorDatabase(mock_account)
-            repl.create_replication(
-                mock_source, mock_target, "REPLID"
-                )
+            repl = Replicator(self.mock_account)
+            repl.create_replication(mock_source, mock_target, "REPLID")
 
         self.assertTrue(mock_create.called)
         repl_doc = mock_create.call_args[0][0]
@@ -99,12 +102,6 @@ class ReplicatorDatabaseTests(unittest.TestCase):
 
     def test_create_replication_errors(self):
         """check expected error conditions"""
-        mock_account = mock.Mock()
-        mock_account.session = mock.Mock()
-        mock_account.session.return_value = {
-            "userCtx": "user Context"
-        }
-
         mock_target = mock.Mock()
         mock_target.database_url = "http://bob.cloudant.com/target"
         mock_target.creds = {'basic_auth': "target_auth"}
@@ -112,7 +109,7 @@ class ReplicatorDatabaseTests(unittest.TestCase):
         mock_source.database_url = "http://bob.cloudant.com/source"
         mock_source.creds = {'basic_auth': "source_auth"}
 
-        repl = ReplicatorDatabase(mock_account)
+        repl = Replicator(self.mock_account)
         self.assertRaises(
             CloudantException,
             repl.create_replication,
@@ -127,34 +124,34 @@ class ReplicatorDatabaseTests(unittest.TestCase):
             )
 
     def test_list_replications(self):
-        with mock.patch('cloudant.database.CloudantDatabase.all_docs') as mock_all_docs:
+        """ test retrieve replications"""
+        with mock.patch('cloudant.database.CouchDatabase.all_docs') as mock_all_docs:
             mock_all_docs.return_value = {
                 "rows": [
-                    {"doc":"replication_1"},
-                    {"doc": "replication_2"}
+                    {"id": "replication_1", "doc": {"_id": "replication_1"}},
+                    {"id": "replication_2", "doc": {"_id": "replication_2"}}
                 ]
             }
-            mock_account = mock.Mock()
-            repl = ReplicatorDatabase(mock_account)
+            repl = Replicator(self.mock_account)
 
             self.assertEqual(
                 repl.list_replications(),
-                ['replication_1', 'replication_2']
+                [{"_id": "replication_1"}, {"_id": "replication_2"}]
             )
+
     def test_replication_state(self):
         """test replication state method"""
-        mock_account = mock.Mock()
-        repl = ReplicatorDatabase(mock_account)
+        repl = Replicator(self.mock_account)
 
         mock_doc = mock.Mock()
         mock_doc.fetch = mock.Mock()
         mock_doc.get = mock.Mock()
         mock_doc.get.return_value = "STATE"
 
-        repl['replication_1'] = mock_doc
+        repl.database['replication_1'] = mock_doc
         self.assertEqual(repl.replication_state('replication_1'), 'STATE')
 
-        with mock.patch('cloudant.replicator.ReplicatorDatabase.__getitem__') as mock_gi:
+        with mock.patch('cloudant.database.CouchDatabase.__getitem__') as mock_gi:
             mock_gi.side_effect = KeyError("womp")
             self.assertRaises(
                 CloudantException,
@@ -164,19 +161,18 @@ class ReplicatorDatabaseTests(unittest.TestCase):
 
     def test_stop_replication(self):
         """test stop_replication call"""
-        mock_account = mock.Mock()
-        repl = ReplicatorDatabase(mock_account)
+        repl = Replicator(self.mock_account)
 
         mock_doc = mock.Mock()
         mock_doc.fetch = mock.Mock()
         mock_doc.delete = mock.Mock()
 
-        repl['replication_1'] = mock_doc
+        repl.database['replication_1'] = mock_doc
         repl.stop_replication('replication_1')
         self.assertTrue(mock_doc.fetch.called)
         self.assertTrue(mock_doc.delete.called)
 
-        with mock.patch('cloudant.replicator.ReplicatorDatabase.__getitem__') as mock_gi:
+        with mock.patch('cloudant.database.CouchDatabase.__getitem__') as mock_gi:
             mock_gi.side_effect = KeyError("womp")
             self.assertRaises(
                 CloudantException,
@@ -186,32 +182,26 @@ class ReplicatorDatabaseTests(unittest.TestCase):
 
     def test_follow_replication(self):
         """test follow replication feature"""
-
-        with mock.patch('cloudant.replicator.ReplicatorDatabase.changes') as mock_changes:
-
+        with mock.patch('cloudant.database.CouchDatabase.changes') as mock_changes:
             mock_changes.return_value = [
                 {"id": "not_this replication"},
                 {"id": "not_this replication"},
                 {"id": "replication_1", "_replication_state": "not finished"},
                 {"id": "replication_1", "_replication_state": "completed"},
             ]
-
-            mock_account = mock.Mock()
-            repl = ReplicatorDatabase(mock_account)
+            repl = Replicator(self.mock_account)
 
             mock_doc = mock.Mock()
             mock_doc.fetch = mock.Mock()
             mock_doc.get = mock.Mock()
             mock_doc.get.side_effect = ['triggered', 'triggered', 'triggered', 'completed']
 
-            repl['replication_1'] = mock_doc
+            repl.database['replication_1'] = mock_doc
 
             for x, i in enumerate(repl.follow_replication('replication_1')):
                 pass
             # expect 4 iterations
             self.assertEqual(x, 3)
-
-
 
 if __name__ == '__main__':
     unittest.main()