+++ CHANGES.rst
@@ -16,6 +16,7 @@
 - [FIX] The Document local dictionary ``_id`` key is now synched with ``_document_id`` private attribute.
 - [FIX] The Document local dictionary is now refreshed after an add/update/delete of an attachment.
 - [FIX] The Document ``fetch()`` method now refreshes the Document local dictionary content correctly.
+- [BREAKING] Replace the ReplicatorDatabase class with the Replicator class.  A Replicator object has a database attribute that represents the _replicator database.  This allows the Replicator to work for both a CloudantDatabase and a CouchDatabase.
 
 2.0.0a1 (2015-10-13)
 ====================